main_server/requirements.txt
fastapi==0.112.2
langchain_community==0.2.14
langchain_core==0.2.36
langchain_openai==0.1.23
langchain_qdrant==0.1.3
langchain_text_splitters==0.2.2
passlib==1.7.4
pika==1.3.2
pydantic==2.8.2
python-dotenv==1.0.1
python_jose==3.3.0
SQLAlchemy==2.0.30
tiktoken==0.7.0
uvicorn==0.30.6
PyMySQL==1.1.1
cryptography==42.0.5
redis==5.0.7
mailjet-rest==1.3.4
Jinja2==3.1.4
python-multipart==0.0.9